requirements.txt
Pillow>=9.5.0
opencv-python==4.5.5.64
numpy>=1.24.3
requests>=2.29.0
tkintermapview @ git+https://github.com/DarkLord76865/TkinterMapView@main
